+++ templates/typescript/src/service_worker/service_worker_controller.d.ts
@@ -1,4 +1,0 @@
-/** Register service worker, if it is supported by browser. */
-export function registerServiceWorker(): void;
-/** Unregister service worker, if it is supported by browser. */
-export function unregisterServiceWorker(): void;